+++ src/Console/ImportStops.php
@@ -3,12 +3,13 @@
 namespace TromsFylkestrafikk\Netex\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use SimpleXMLElement;
 use TromsFylkestrafikk\Netex\Models\StopPlace;
-use TromsFylkestrafikk\Xml\ChristmasTreeParser;
 use TromsFylkestrafikk\Netex\Services\StopsActivator;
+use TromsFylkestrafikk\Xml\ChristmasTreeParser;
 
 class ImportStops extends Command
 {
@@ -36,17 +37,27 @@ class ImportStops extends Command
     protected $progressBar = null;
 
     /**
+     * Number of elements to parse in XML.
+     *
+     * @var int
+     */
+    protected $xmlElements = 0;
+
+    /**
      * Import statistics.
      */
     protected $processedStops = 0;
     protected $processedGroups = 0;
     protected $processedPlaces = 0;
+    protected $processedTariffs = 0;
     protected $updatedStops = 0;
     protected $updatedGroups = 0;
     protected $updatedPlaces = 0;
+    protected $updatedTariffs = 0;
     protected $createdStops = 0;
     protected $createdGroups = 0;
     protected $createdPlaces = 0;
+    protected $createdTariffs = 0;
 
     /**
      * Collections of items not seen during import, keyed by model name.
@@ -79,23 +90,17 @@ class ImportStops extends Command
     public function handle(StopsActivator $activator)
     {
         $xmlFile = $this->argument('xml');
-        $this->logInfo("netex:importstops: BEGIN: %s", $xmlFile);
+        $this->logInfo("[netex:importstops] BEGIN: %s", $xmlFile);
+        $this->scanXml($xmlFile);
         $reader = new ChristmasTreeParser();
-        if (!$reader->open($xmlFile)) {
-            $this->error("Could not open XML file for reading: '$xmlFile'");
-        }
+        $reader->open($xmlFile);
         $this->info("Importing ...");
         $this->unseen['StopPlace'] = DB::table('netex_stop_place')->pluck('id', 'id');
         $this->unseen['StopQuay'] = DB::table('netex_stop_quay')->pluck('id', 'id');
         $this->unseen['GroupOfStopPlaces'] = DB::table('netex_group_of_stop_places')->pluck('id', 'id');
         $this->unseen['TopographicPlace'] = DB::table('netex_topographic_place')->pluck('id', 'id');
         $this->unseen['TariffZone'] = DB::table('netex_tariff_zone')->pluck('id', 'id');
-        $this->progressBar = $this->output->createProgressBar(array_reduce(
-            ['StopPlace', 'GroupOfStopPlaces', 'TopographicPlace', 'TariffZone'],
-            function ($carry, $item) {
-                return $carry +  $this->unseen[$item]->count();
-            }
-        ));
+        $this->progressBar = $this->output->createProgressBar($this->xmlElements);
         $this->progressBar->start();
         $reader
             ->addCallback(['SiteFrame', 'stopPlaces', 'StopPlace'], [$this, 'readStopPlace'])
@@ -113,6 +118,7 @@ class ImportStops extends Command
             $this->deleteOld();
         }
         $this->importSummary();
+        return self::SUCCESS;
     }
 
     /**
@@ -245,6 +251,8 @@ class ImportStops extends Command
 
     public function readTariffZone(ChristmasTreeParser $reader)
     {
+        $this->processedTariffs++;
+        $this->progressBar->advance();
         $xml = $reader->expandSimpleXml();
         /** @var \TromsFylkestrafikk\Netex\Models\TariffZone|null $tzone */
         $tzone = $this->prepareNetexModel($xml, 'TariffZone');
@@ -257,6 +265,7 @@ class ImportStops extends Command
         $tzone->validFromDate = $xml->ValidBetween->FromDate;
         $tzone->validToDate = $xml->ValidBetween->ToDate;
         $tzone->polygon_poslist = $this->getPolygonPoslist($xml);
+        $tzone->updated_at ? $this->updatedTariffs++ : $this->createdTariffs++;
 
         self::nullifyObjectProps($tzone, ['validFromDate', 'validToDate']);
         $tzone->save();
@@ -361,6 +370,29 @@ class ImportStops extends Command
         DB::table('netex_stop_quay_alt_id')->whereIn('stop_quay_id', $this->unseen['StopQuay']->toArray())->delete();
     }
 
+    /**
+     * Scan XML quickly and count targeted elements.
+     *
+     * @param string $xmlFile Input XML file
+     *
+     * @return int
+     */
+    protected function scanXml($xmlFile)
+    {
+        $reader = new ChristmasTreeParser();
+        $reader->open($xmlFile);
+        $counter = fn () => $this->xmlElements++;
+        $reader
+            ->addCallback(['SiteFrame', 'stopPlaces', 'StopPlace'], $counter)
+            ->addCallback(['SiteFrame', 'groupsOfStopPlaces', 'GroupOfStopPlaces'], $counter)
+            ->addCallback(['SiteFrame', 'topographicPlaces', 'TopographicPlace'], $counter)
+            ->addCallback(['SiteFrame', 'tariffZones', 'TariffZone'], $counter)
+            ->parse()
+            ->close();
+        $this->info(sprintf("Found %d elements", $this->xmlElements));
+        return $this->xmlElements;
+    }
+
     protected function importSummary()
     {
         $this->info(sprintf(
@@ -368,6 +400,12 @@ class ImportStops extends Command
             $this->processedPlaces,
             $this->updatedPlaces,
             $this->createdPlaces
+        ));
+        $this->info(sprintf(
+            "TaroffZone: %d processed, %d updated, %d created",
+            $this->processedTariffs,
+            $this->updatedTariffs,
+            $this->createdTariffs
         ));
         $this->info(sprintf(
             "GroupOfStopPlaces: %d processed, %d updated, %d created",
@@ -383,18 +421,22 @@ class ImportStops extends Command
         ));
         if (!$this->option('keep')) {
             $this->info(sprintf(
-                "DELETIONS: StopPlace: %d, StopQuay: %d, GroupOfStopPlaces: %d, TopographicPlace: %d",
+                "DELETIONS: StopPlace: %d, StopQuay: %d, GroupOfStopPlaces: %d, TariffPlace: %d, TopographicPlace: %d",
                 $this->unseen['StopPlace']->count(),
                 $this->unseen['StopQuay']->count(),
                 $this->unseen['GroupOfStopPlaces']->count(),
+                $this->unseen['TariffZone']->count(),
                 $this->unseen['TopographicPlace']->count(),
             ));
         }
         $this->logInfo(
-            "netex:importstops: END. Imported %d stops. %d new, %d updated.",
+            "[netex:importstops] END. Imported %d stops. %d new, %d updated. Processed %d groups of stop places, %d tariff zones and %d topographic places",
             $this->processedStops,
             $this->createdStops,
-            $this->updatedStops
+            $this->updatedStops,
+            $this->processedGroups,
+            $this->processedTariffs,
+            $this->processedPlaces
         );
     }